multiapp (fanout) distribution fix for more than 32 queues

+++ userland/examples_zc/zbalance_ipc.c
@@ -570,7 +570,7 @@ int64_t fo_multiapp_ip_distribution_func(pfring_zc_pkt_buff *pkt_handle, pfring_
 
   for (i = 0; i < num_apps; i++) {
     app_instance = hash % instances_per_app[i];
-    consumers_mask |= (1 << (offset + app_instance));
+    consumers_mask |= ((int64_t) 1 << (offset + app_instance));
     offset += instances_per_app[i];
   }
